@@ -234,8 +234,10 @@
     "@faker-js/faker": "9.2.0",
     "@ffmpeg-installer/ffmpeg": "1.1.0",
     "@types/chai": "4.3.6",
+    "@types/copy-webpack-plugin": "8.0.0",
     "@types/express": "4.17.17",
     "@types/i18next-fs-backend": "1.1.2",
+    "@types/js-yaml": "4.0.9",
     "@types/lodash": "4.14.192",
     "@types/mocha": "10.0.1",
     "@types/mock-fs": "4.13.4",
@@ -309,10 +311,10 @@
     "ts-mocha": "10.0.0",
     "ts-node": "10.9.1",
     "typescript": "5.3.3",
-    "webpack": "^5.96",
-    "webpack-assets-manifest": "5.1.0",
-    "webpack-cli": "5.0.1",
-    "webpack-merge": "5.8.0"
+    "webpack": "^5.100.2",
+    "webpack-assets-manifest": "6.2.1",
+    "webpack-cli": "6.0.1",
+    "webpack-merge": "6.0.1"
   },
   "engines": {
     "node": "22",
